Extract Core project
