+++ appveyor.yml
@@ -1,4 +1,4 @@
-version: 1.2.{build}
+version: 1.3.{build}
 image: Visual Studio 2019
 configuration: Release
 dotnet_csproj:
