+++ src/SmartCodeGenerator.Core/GeneratedDocument.cs
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Simplification;
 using SmartCodeGenerator.Sdk;
 
-namespace SmartCodeGenerator.Engine
+namespace SmartCodeGenerator.Core
 {
     internal class GeneratedDocument
     {
