+++ src/SmartCodeGenerator.Core/GeneratorPluginProvider.cs
@@ -5,7 +5,7 @@ using System.Reflection;
 using Microsoft.CodeAnalysis;
 using SmartCodeGenerator.Sdk;
 
-namespace SmartCodeGenerator.Engine
+namespace SmartCodeGenerator.Core
 {
     public class GeneratorPluginProvider
     {
