+++ src/SmartCodeGenerator.Core/ITransformedDocumentPersister.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace SmartCodeGenerator.Engine
+namespace SmartCodeGenerator.Core
 {
     public interface ITransformedDocumentPersister
     {
