+++ src/SmartCodeGenerator.Core/ParallelHelper.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace SmartCodeGenerator.Engine
+namespace SmartCodeGenerator.Core
 {
     internal static class ParallelHelper
     {
