+++ src/SmartCodeGenerator.Core/SmartCodeGenerator.Core.csproj
@@ -3,14 +3,16 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
+
   <ItemGroup>
-    <PackageReference Include="ApprovalTests" Version="4.2.0" />
-    <PackageReference Include="DiffPlex" Version="1.5.0" />
+    <ProjectReference Include="..\SmartCodeGenerator.Sdk\SmartCodeGenerator.Sdk.csproj" />
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="2.9.5" PrivateAssets="all" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="3.3.1" />
     <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Workspaces" Version="3.3.1" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Workspaces.MSBuild" Version="3.3.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
   </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\SmartCodeGenerator.Core\SmartCodeGenerator.Core.csproj" />
-  </ItemGroup>
+
 </Project>
