+++ src/SmartCodeGenerator.Engine/MsBuildHelper.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Build.Locator;
 using Microsoft.CodeAnalysis.MSBuild;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.Engine
 {
