+++ src/SmartCodeGenerator.Engine/OptionsHelper.cs
@@ -1,5 +1,6 @@
 using System;
 using CommandLine;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.Engine
 {
