+++ src/SmartCodeGenerator.Engine/Program.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Build.Locator;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.Engine
 {
