+++ src/SmartCodeGenerator.Engine/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "SmartCodeGenerator": {
       "commandName": "Project",
-      "commandLineArgs": "-p \"c:\\Repositories\\MappingGenerator\\MappingGenerator\\OnBuildGenerator.Sample\\OnBuildGenerator.Sample.csproj\" -o \"c:\\tmp\\generator\" -g \"c:\\Repositories\\MappingGenerator\\MappingGenerator\\OnBuildGenerator\\bin\\Debug\\netcoreapp3.0\\OnBuildGenerator.dll\""
+      "commandLineArgs": "-p \"c:\\Users\\Cezar\\source\\repos\\ConsoleApp11\\ConsoleApp11\\ConsoleApp11.csproj\" -o \"c:\\tmp\\generator\" -g \"c:\\Repositories\\MappingGenerator\\MappingGenerator\\OnBuildGenerator\\bin\\Debug\\netcoreapp3.0\\OnBuildGenerator.dll\""
     }
   }
 }
