+++ src/SmartCodeGenerator.Engine/ToolsOptionsExtensions.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.Engine
 {
