+++ src/SmartCodeGenerator.Engine/TransformedDocumentPersister.cs
@@ -5,6 +5,7 @@ using System.Security.Cryptography;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.Engine
 {
