+++ src/SmartCodeGenerator.TestKit/InMemoryDocumentPersister.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using SmartCodeGenerator.Engine;
+using SmartCodeGenerator.Core;
 
 namespace SmartCodeGenerator.TestKit
 {
